separate cf compliant nc writer functions from make_grids
