+++ grid/make_grids.py
@@ -4,10 +4,7 @@ import glob
 import os, sys
 from datetime import datetime, timedelta
 
-from math import ceil
-
 import numpy as np
-import tables
 
 from . import density_to_files
 from lmatools.stream.subset import broadcast
@@ -15,6 +12,8 @@ from lmatools.io.LMA_h5_file import read_flashes, to_seconds
 
 from lmatools.coordinateSystems import MapProjection, GeographicSystem
 from six.moves import range
+
+from .cf_netcdf import write_cf_netcdf, write_cf_netcdf_3d, write_cf_netcdf_latlon, write_cf_netcdf_3d_latlon
     
 
 def dlonlat_at_grid_center(ctr_lat, ctr_lon, dx=4.0e3, dy=4.0e3,
@@ -67,299 +66,6 @@ def dlonlat_at_grid_center(ctr_lat, ctr_lon, dx=4.0e3, dy=4.0e3,
     
     return dlon, dlat, (lon_min, lon_max), (lat_min, lat_max)
     
-
-def write_cf_netcdf_latlon(outfile, t_start, t, xloc, yloc, lon_for_x, lat_for_y, ctr_lat, ctr_lon, grid, grid_var_name, grid_description, format='i', **kwargs):
-    """ Write a Climate and Forecast Metadata-compliant NetCDF file. 
-        
-        Grid is regular in lon, lat and so no map projection information is necessary.
-    
-        Should display natively in conformant packages like McIDAS-V.
-        
-    """
-    try:
-        from netCDF4 import Dataset as NetCDFFile
-    except ImportError:
-        from scipy.io.netcdf import NetCDFFile
-
-    missing_value = -9999
-    
-    nc_out = NetCDFFile(outfile, 'w')
-    nc_out.createDimension('lon', xloc.shape[0])
-    nc_out.createDimension('lat', yloc.shape[0])
-    nc_out.createDimension('ntimes', t.shape[0])  #unlimited==None
-
-    # declare the coordinate reference system, WGS84 values
-    proj = nc_out.createVariable('crs', 'i', ())
-    proj.grid_mapping_name = 'latitude_longitude'
-    proj.longitude_of_prime_meridian = 0.0 
-    proj.semi_major_axis = 6378137.0 
-    proj.inverse_flattening = 298.257223563 
-    
-    y_coord = nc_out.createVariable('latitude', 'f', ('lat',))
-    y_coord.units = "degrees_north"
-    y_coord.long_name = "latitude"
-    y_coord.standard_name = 'latitude'
-
-    x_coord = nc_out.createVariable('longitude', 'f', ('lon',))
-    x_coord.units = "degrees_east"
-    x_coord.long_name = "longitude"
-    x_coord.standard_name = 'longitude'
-
-    times = nc_out.createVariable('time', 'f', ('ntimes',) )#, filters=no_compress)
-    times.long_name="time"
-    times.units = "seconds since %s" % t_start.strftime('%Y-%m-%d %H:%M:%S')
-    
-    # lons = nc_out.createVariable('lons', 'd', ('nx','ny') )#, filters=no_compress)
-    # lons.long_name="longitude"
-    # lons.standard_name="longitude"
-    # lons.units = "degrees_east"
-    # 
-    # lats = nc_out.createVariable('lats', 'd', ('nx','ny') )#, filters=no_compress)
-    # lats.long_name="latitude"
-    # lats.standard_name="latitude"
-    # lats.units = "degrees_north"
-
-    lightning2d = nc_out.createVariable(grid_var_name, format, ('ntimes','lon','lat'),  zlib=True)#, filters=no_compress)
-    lightning2d.long_name=grid_description #'LMA VHF event counts (vertically integrated)'
-    lightning2d.units=kwargs['grid_units']
-    # lightning2d.coordinates='time lons lats'
-    lightning2d.grid_mapping = "crs"
-    lightning2d.missing_value = missing_value
-
-    x_coord[:] = xloc[:]
-    y_coord[:] = yloc[:]
-    times[:] = t[:]
-    # lons[:] = lon_for_x[:]
-    # lats[:] = lat_for_y[:]
-
-    for i in range(grid.shape[2]):
-        lightning2d[i,:,:] = grid[:,:,i]
-    nc_out.close()    
-
-def write_cf_netcdf(outfile, t_start, t, xloc, yloc, lon_for_x, lat_for_y, ctr_lat, ctr_lon, grid, grid_var_name, grid_description, format='i', **kwargs):
-    """ Write a Climate and Forecast Metadata-compliant NetCDF file. 
-    
-        Should display natively in conformant packages like McIDAS-V.
-        
-    """
-
-    import scipy.io.netcdf as nc
-
-    missing_value = -9999
-    
-    nc_out = nc.NetCDFFile(outfile, 'w')
-    nc_out.createDimension('nx', xloc.shape[0])
-    nc_out.createDimension('ny', yloc.shape[0])
-    nc_out.createDimension('ntimes', t.shape[0])  #unlimited==None
-
-    proj = nc_out.createVariable('Lambert_Azimuthal_Equal_Area', 'i', ())
-    proj.grid_mapping_name = 'lambert_azimuthal_equal_area'
-    proj.longitude_of_projection_origin = ctr_lon
-    proj.latitude_of_projection_origin = ctr_lat
-    proj.false_easting = 0.0
-    proj.false_northing = 0.0
-    
-    # x_coord = nc_out.createVariable('longitude', 'f', ('nx',))
-    # x_coord.long_name="longitude"
-    # x_coord.standard_name="longitude"
-    # x_coord.units = "degrees_east"
-    x_coord = nc_out.createVariable('x', 'f', ('nx',))
-    x_coord.units = "km"
-    x_coord.long_name = "x coordinate of projection"
-    x_coord.standard_name = 'projection_x_coordinate'
-
-    # y_coord = nc_out.createVariable('latitude', 'f', ('nx',))
-    # y_coord.long_name="latitude"
-    # y_coord.standard_name="latitude"
-    # y_coord.units = "degrees_north"
-    y_coord = nc_out.createVariable('y', 'f', ('ny',))
-    y_coord.units = "km"
-    y_coord.long_name = "y coordinate of projection"
-    y_coord.standard_name = 'projection_y_coordinate'
-
-    times = nc_out.createVariable('time', 'f', ('ntimes',) )#, filters=no_compress)
-    times.long_name="time"
-    times.units = "seconds since %s" % t_start.strftime('%Y-%m-%d %H:%M:%S')
-    
-    #Dtype change from 'd' to 'f':
-    lons = nc_out.createVariable('lons', 'f', ('nx','ny') )#, filters=no_compress)
-    lons.long_name="longitude"
-    lons.standard_name="longitude"
-    lons.units = "degrees_east"
-    
-    lats = nc_out.createVariable('lats', 'f', ('nx','ny') )#, filters=no_compress)
-    lats.long_name="latitude"
-    lats.standard_name="latitude"
-    lats.units = "degrees_north"
-
-    lightning2d = nc_out.createVariable(grid_var_name, format, ('ntimes','nx','ny') )#, filters=no_compress)
-    lightning2d.long_name=grid_description #'LMA VHF event counts (vertically integrated)'
-    lightning2d.units='dimensionless'
-    lightning2d.coordinates='time lons lats'
-    lightning2d.grid_mapping = "Lambert_Azimuthal_Equal_Area"
-    lightning2d.missing_value = missing_value
-
-    x_coord[:] = xloc[:]
-    y_coord[:] = yloc[:]
-    times[:] = t[:]
-    lons[:] = lon_for_x[:]
-    lats[:] = lat_for_y[:]
-
-    for i in range(grid.shape[2]):
-        lightning2d[i,:,:] = grid[:,:,i]
-    nc_out.close()
-
-def write_cf_netcdf_3d(outfile, t_start, t, xloc, yloc, zloc, lon_for_x, lat_for_y, alt_for_z, ctr_lat, ctr_lon, ctr_alt, grid, grid_var_name, grid_description, format='i', **kwargs):
-    """ Write a Climate and Forecast Metadata-compliant NetCDF file. 
-    
-        Should display natively in conformant packages like McIDAS-V.
-        
-    """
-
-    import scipy.io.netcdf as nc
-
-    missing_value = -9999
-    
-    nc_out = nc.NetCDFFile(outfile, 'w')
-    nc_out.createDimension('nx', xloc.shape[0])
-    nc_out.createDimension('ny', yloc.shape[0])
-    nc_out.createDimension('nz', zloc.shape[0])
-    nc_out.createDimension('ntimes', t.shape[0])  #unlimited==None
-
-    proj = nc_out.createVariable('Lambert_Azimuthal_Equal_Area', 'i', ())
-    proj.grid_mapping_name = 'lambert_azimuthal_equal_area'
-    proj.longitude_of_projection_origin = ctr_lon
-    proj.latitude_of_projection_origin = ctr_lat
-    proj.altitude_of_projection_origin = ctr_alt
-    proj.false_easting = 0.0
-    proj.false_northing = 0.0
-    
-    # x_coord = nc_out.createVariable('longitude', 'f', ('nx',))
-    # x_coord.long_name="longitude"
-    # x_coord.standard_name="longitude"
-    # x_coord.units = "degrees_east"
-    x_coord = nc_out.createVariable('x', 'f', ('nx',))
-    x_coord.units = "km"
-    x_coord.long_name = "x coordinate of projection"
-    x_coord.standard_name = 'projection_x_coordinate'
-
-    # y_coord = nc_out.createVariable('latitude', 'f', ('nx',))
-    # y_coord.long_name="latitude"
-    # y_coord.standard_name="latitude"
-    # y_coord.units = "degrees_north"
-    y_coord = nc_out.createVariable('y', 'f', ('ny',))
-    y_coord.units = "km"
-    y_coord.long_name = "y coordinate of projection"
-    y_coord.standard_name = 'projection_y_coordinate'
-
-    z_coord = nc_out.createVariable('z', 'f', ('nz',))
-    z_coord.units = 'km'
-    z_coord.long_name = "z coordinate of projection"
-    z_coord.standard_name = 'projection_z_coordinate'
-
-    times = nc_out.createVariable('time', 'f', ('ntimes',) )#, filters=no_compress)
-    times.long_name="time"
-    times.units = "seconds since %s" % t_start.strftime('%Y-%m-%d %H:%M:%S')
-    
-    #Dtype changed due to previous issue with pupynere.
-    #---------------------------------------------------
-    lons = nc_out.createVariable('lons', 'f', ('nx','ny','nz') )#, filters=no_compress)
-    lons.long_name="longitude"
-    lons.standard_name="longitude"
-    lons.units = "degrees_east"
-    
-    lats = nc_out.createVariable('lats', 'f', ('nx','ny', 'nz') )#, filters=no_compress)
-    lats.long_name="latitude"
-    lats.standard_name="latitude"
-    lats.units = "degrees_north"
-    
-    alts = nc_out.createVariable('alts', 'f', ('nx','ny', 'nz') )#, filters=no_compress)
-    alts.long_name="altitude"
-    alts.standard_name="altitude"
-    alts.units = "meters"
-
-    lightning3d = nc_out.createVariable(grid_var_name, format, ('ntimes','nx','ny', 'nz') )#, filters=no_compress)
-    lightning3d.long_name=grid_description #'LMA VHF event counts (vertically integrated)'
-    lightning3d.units='dimensionless'
-    lightning3d.coordinates='time lons lats alts'
-    lightning3d.grid_mapping = "Lambert_Azimuthal_Equal_Area"
-    lightning3d.missing_value = missing_value
-
-    x_coord[:] = xloc[:]
-    y_coord[:] = yloc[:]
-    z_coord[:] = zloc[:]
-    times[:] = t[:]
-    lons[:] = lon_for_x[:]
-    lats[:] = lat_for_y[:]
-    alts[:] = alt_for_z[:]
-
-    for i in range(grid.shape[3]):
-        lightning3d[i,:,:,:] = grid[:,:,:,i]
-    nc_out.close()
-
-def write_cf_netcdf_3d_latlon(outfile, t_start, t, xloc, yloc, zloc, lon_for_x, lat_for_y, alt_for_z, ctr_lat, ctr_lon, ctr_alt, grid, grid_var_name, grid_description, format='i', **kwargs):
-    """ Write a Climate and Forecast Metadata-compliant NetCDF file. 
-    
-        Should display natively in conformant packages like McIDAS-V.
-        
-    """
-
-    import scipy.io.netcdf as nc
-
-    missing_value = -9999
-    
-    nc_out = nc.NetCDFFile(outfile, 'w')
-    nc_out.createDimension('lon', xloc.shape[0])
-    nc_out.createDimension('lat', yloc.shape[0])
-    nc_out.createDimension('alt', zloc.shape[0])
-    nc_out.createDimension('ntimes', t.shape[0])  #unlimited==None
-
-    # declare the coordinate reference system, WGS84 values
-    proj = nc_out.createVariable('crs', 'i', ())
-    proj.grid_mapping_name = 'latitude_longitude'
-    proj.longitude_of_prime_meridian = 0.0 
-    proj.semi_major_axis = 6378137.0 
-    proj.inverse_flattening = 298.257223563 
-    
-    y_coord = nc_out.createVariable('latitude', 'f', ('lat',))
-    y_coord.units = "degrees_north"
-    y_coord.long_name = "latitude"
-    y_coord.standard_name = 'latitude'
-
-    x_coord = nc_out.createVariable('longitude', 'f', ('lon',))
-    x_coord.units = "degrees_east"
-    x_coord.long_name = "longitude"
-    x_coord.standard_name = 'longitude'
-
-    z_coord = nc_out.createVariable('altitude', 'f', ('alt',))
-    z_coord.units = 'km'
-    z_coord.long_name = "height above mean sea level"
-    z_coord.standard_name = 'altitude'
-    z_coord.positive = 'up'
-
-    times = nc_out.createVariable('time', 'f', ('ntimes',) )#, filters=no_compress)
-    times.long_name="time"
-    times.units = "seconds since %s" % t_start.strftime('%Y-%m-%d %H:%M:%S')
-    
-    lightning3d = nc_out.createVariable(grid_var_name, format, ('ntimes','lon','lat','alt') )#, filters=no_compress)
-    lightning3d.long_name=grid_description #'LMA VHF event counts (vertically integrated)'
-    lightning3d.units=kwargs['grid_units']
-    # lightning3d.coordinates='time lons lats alts'
-    lightning3d.grid_mapping = "crs"
-    lightning3d.missing_value = missing_value
-
-    x_coord[:] = xloc[:]
-    y_coord[:] = yloc[:]
-    z_coord[:] = zloc[:]
-    times[:] = t[:]
-    # lons[:] = lon_for_x[:]
-    # lats[:] = lat_for_y[:]
-    # alts[:] = alt_for_z[:]
-
-    for i in range(grid.shape[3]):
-        lightning3d[i,:,:,:] = grid[:,:,:,i]
-    nc_out.close()
-
 
 def time_edges(start_time, end_time, frame_interval):
     """ Return lists cooresponding the start and end times of frames lasting frame_interval
@@ -419,7 +125,7 @@ class FlashGridder(object):
         self.frame_interval = frame_interval
         self.base_date = base_date
         self.min_points_per_flash = 1
-        self.proj_name = 'aeqd'
+        self.proj_name = proj_name
         self.ctr_lat, self.ctr_lon, self.ctr_alt = ctr_lat, ctr_lon, ctr_alt
         
         if flash_count_logfile is None:
@@ -441,12 +147,12 @@ class FlashGridder(object):
         y0 = yedge[0]
         z0 = zedge[0]
     
-        if proj_name == 'latlong':
+        if self.proj_name == 'latlong':
             dx_units = '{0:6.4f}deg'.format(dx)
             mapProj = GeographicSystem()
         else:
             dx_units = '{0:5.1f}m'.format(dx)
-            mapProj = MapProjection(projection=proj_name, ctrLat=ctr_lat, ctrLon=ctr_lon, lat_ts=ctr_lat, 
+            mapProj = MapProjection(projection=self.proj_name, ctrLat=ctr_lat, ctrLon=ctr_lon, lat_ts=ctr_lat, 
                                 lon_0=ctr_lon, lat_0=ctr_lat, lat_1=ctr_lat, ellipse=proj_ellipse, datum=proj_datum)
         geoProj = GeographicSystem()
